feat: add stuck protection backtrack and example

+++ TeamCode/src/main/java/org/robotics/robotics/xdk/teamcode/autonomous/position/RobotStuckProtection.kt
@@ -1,7 +1,7 @@
 package org.robotics.robotics.xdk.teamcode.autonomous.position
 
 data class RobotStuckProtection(
-    val minimumRequiredTranslationalDifference: Int,
-    val minimumRequiredRotationalDifference: Int,
+    val minimumRequiredTranslationalDifference: Double,
+    val minimumRequiredRotationalDifference: Double,
     val minimumMillisUntilDeemedStuck: Long
 )
